+++ application/eu-central-1/env/variables.tf
@@ -6,6 +6,11 @@ variable "k8s_private_subnets" {
   type = "list"
 }
 
+variable "k8s_public_subnets" {
+  type    = "list"
+  default = []
+}
+
 variable "k8s_node_count" {}
 
 variable "region" {}
@@ -33,3 +38,15 @@ variable "k8s_node_instance_type" {}
 variable "k8s_aws_ssh_keypair_name" {}
 
 variable "k8s_linux_distro" {}
+
+variable "k8s_enable_pod_autoscaling" {}
+
+variable "k8s_enable_cluster_autoscaling" {}
+
+variable "k8s_protect_cluster_from_scale_in" {}
+
+variable "k8s_install_helm" {}
+
+variable "k8s_allowed_worker_ssh_cidrs" {
+  type = "list"
+}